Move ArrayIterator into its own header for reuse
